refactor(openfeature): extract DatadogCoreOpenFeatureProvider base
Both the online and offline providers now extend a shared, internal (un-exported)
DatadogCoreOpenFeatureProvider that owns the FlagsClient, event emitter, and
resolveX evaluation. DatadogOpenFeatureProvider keeps its fetch-on-context
initialize/onContextChange, and DatadogOfflineOpenFeatureProvider extends the
base directly instead of the online provider, so it no longer inherits the
online fetch/promise-chain machinery. Public API and online behavior are
unchanged and the base is not exported. Matches the Portable RFC's CoreProvider.

Adds a provider test locking the online provider's fetch-on-initialize behavior.
