+++ packages/react-native-openfeature/src/provider.ts
@@ -4,58 +4,25 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-import { DdFlags } from '@datadog/mobile-react-native';
-import type {
-    FlagDetails,
-    FlagsClient,
-    EvaluationContext as DdEvaluationContext
-} from '@datadog/mobile-react-native';
-import { OpenFeatureEventEmitter, ErrorCode } from '@openfeature/web-sdk';
 import type {
     EvaluationContext as OFEvaluationContext,
-    JsonValue,
-    Logger,
-    Paradigm,
-    Provider,
-    ProviderMetadata,
-    ResolutionDetails,
-    PrimitiveValue,
-    ProviderEventEmitter,
-    ProviderEvents
+    ProviderMetadata
 } from '@openfeature/web-sdk';
 
-export interface DatadogOpenFeatureProviderOptions {
-    /**
-     * The name of the Datadog Flags client to use.
-     *
-     * Provide this parameter in order to use different Datadog Flags clients for different OpenFeature domains.
-     *
-     * @default 'default'
-     */
-    clientName?: string;
-}
+import { DatadogCoreOpenFeatureProvider, toDdContext } from './coreProvider';
 
-export class DatadogOpenFeatureProvider implements Provider {
-    readonly runsOn: Paradigm = 'client';
+export type { DatadogOpenFeatureProviderOptions } from './coreProvider';
+
+/**
+ * The online Datadog OpenFeature provider. Fetches precomputed flag assignments from Datadog
+ * whenever the evaluation context is set or changed.
+ */
+export class DatadogOpenFeatureProvider extends DatadogCoreOpenFeatureProvider {
     readonly metadata: ProviderMetadata = {
         name: 'datadog-react-native'
     };
 
-    private options: DatadogOpenFeatureProviderOptions;
-    protected flagsClient: FlagsClient;
-
-    readonly events: ProviderEventEmitter<ProviderEvents> = new OpenFeatureEventEmitter();
     private contextChangePromise = Promise.resolve();
-
-    constructor(options: DatadogOpenFeatureProviderOptions = {}) {
-        if (!options.clientName) {
-            options.clientName = 'default';
-        }
-
-        this.options = options;
-
-        this.flagsClient = DdFlags.getClient(this.options.clientName);
-    }
 
     async initialize(context: OFEvaluationContext = {}): Promise<void> {
         const ddContext = toDdContext(context);
@@ -80,103 +47,4 @@ export class DatadogOpenFeatureProvider implements Provider {
         // Wait for the current context change to complete.
         await this.contextChangePromise;
     }
-
-    resolveBooleanEvaluation(
-        flagKey: string,
-        defaultValue: boolean,
-        _context: OFEvaluationContext,
-        _logger: Logger
-    ): ResolutionDetails<boolean> {
-        const details = this.flagsClient.getBooleanDetails(
-            flagKey,
-            defaultValue
-        );
-        return toFlagResolution(details);
-    }
-
-    resolveStringEvaluation(
-        flagKey: string,
-        defaultValue: string,
-        _context: OFEvaluationContext,
-        _logger: Logger
-    ): ResolutionDetails<string> {
-        const details = this.flagsClient.getStringDetails(
-            flagKey,
-            defaultValue
-        );
-        return toFlagResolution(details);
-    }
-
-    resolveNumberEvaluation(
-        flagKey: string,
-        defaultValue: number,
-        _context: OFEvaluationContext,
-        _logger: Logger
-    ): ResolutionDetails<number> {
-        const details = this.flagsClient.getNumberDetails(
-            flagKey,
-            defaultValue
-        );
-        return toFlagResolution(details);
-    }
-
-    resolveObjectEvaluation<T extends JsonValue>(
-        flagKey: string,
-        defaultValue: T,
-        _context: OFEvaluationContext,
-        _logger: Logger
-    ): ResolutionDetails<T> {
-        // The OpenFeature spec states that the return value can be any valid JSON value.
-        // However, the Datadog Flags feature only supports JSON objects for the JSON feature flag type.
-        // Thus, the user should always expect the returned value to be an object instead of any arbitrary JSON value.
-        // Also, the user is responsible for providing a proper `defaultValue` that's an object.
-
-        const details = this.flagsClient.getObjectDetails<T>(
-            flagKey,
-            defaultValue
-        );
-        return toFlagResolution(details);
-    }
 }
-
-export const toDdContext = (
-    context: OFEvaluationContext
-): DdEvaluationContext => {
-    const { targetingKey, ...attributes } = context;
-
-    // Important ⚠️
-    // The Flags SDK doesn't support nested non-primitive values in the evaluation context as per OF.3 FFE SDK requirement.
-    // However, we let the SDK handle this inside of FlagsClient since it does this processing anyways.
-    const ddContextAttributes = attributes as Record<string, PrimitiveValue>;
-
-    return {
-        // Allow flag evaluations without a provided targeting key.
-        targetingKey: targetingKey ?? '',
-        attributes: ddContextAttributes
-    };
-};
-
-const toFlagResolution = <T>(details: FlagDetails<T>): ResolutionDetails<T> => {
-    const {
-        value,
-        reason,
-        variant,
-        allocationKey,
-        errorCode,
-        errorMessage
-    } = details;
-
-    const parsedErrorCode =
-        errorCode && (ErrorCode[errorCode as ErrorCode] || ErrorCode.GENERAL);
-
-    const result: ResolutionDetails<T> = {
-        value,
-        reason,
-        variant,
-        flagMetadata: allocationKey ? { allocationKey } : undefined,
-        errorCode: parsedErrorCode,
-        errorMessage
-    };
-
-    return result;
-};
